project/BNFC.scala: fix BNFC-related build problems
Previously, when users cleaned BNFC-generated files and then attempted
to generate them again, sbt would first attempt to compile files which
depended on the generated code and fail.

This problem has been addressed by introducing a 'bnfc' configuration to
the sbt build, which sets the 'scalaSource' directory to the value of
'javaSource' to prevent this problem from occurring.

The commands for using BNFC have also changed.  They are now as follows:

  bnfc:generate

  bnfc:clean

  bnfc:generateDocs

  bnfc:cleanDocs